requirements.txt
# =====================================================
# AI Video Summarizer - Production Dependencies
# =====================================================

# --- Web Framework ---
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# --- Database ---
pymongo>=4.6.0
motor>=3.3.0

# --- AI / ML Core ---
torch>=2.2.0
transformers>=4.36.0
sentence-transformers>=2.2.2
openai-whisper>=20231117

# --- Vector Search ---
faiss-cpu>=1.7.4

# --- Transcript Fetching ---
youtube-transcript-api>=1.0.0
yt-dlp>=2024.1.0

# --- Voice (FREE) ---
edge-tts>=6.1.0

# --- Auth & Security ---
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# --- Utilities ---
numpy>=1.24.0
scikit-learn>=1.3.0
aiohttp>=3.9.0
requests>=2.31.0
tqdm>=4.66.0
regex>=2024.1.0

# --- NLP ---
nltk>=3.8.0
sentencepiece>=0.2.0

# --- Gemini AI ---
google-genai>=1.0.0

# --- Export (PDF) ---
fpdf2>=2.7.0

# --- Testing ---
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
